Fix compilation
Conflicts:
	src/main.cpp

+++ src/tp1/Complexe.h
@@ -1,6 +1,6 @@
 struct Complexe {
-    int re;
-    int im;
+    float re;
+    float im;
 };
 
 Complexe addition(Complexe& c1, Complexe& c2);
